Booking tables use fk to timetables table Remove huor column and fix queries

+++ src/infrastructure/dao/bookingDao.ts
@@ -13,36 +13,46 @@ import {
   IBookingRepository,
 } from "../../application/booking/index.js";
 import { buildSortBy } from "./utils.js";
+import { NotFoundException } from "../../application/commons/exceptions.js";
 
 export class BookingDao implements IBookingRepository {
   constructor(protected readonly db: Kysely<DB>) {}
 
   protected readonly DEFAULT_SELECT_FIELDS = [
-    "id",
-    "fullname",
-    "mail",
-    "day",
-    "hour",
-    "createdAt",
-    "updatedAt",
-  ] satisfies ReadonlyArray<SelectExpression<DB, "bookings">>;
+    "bookings.id",
+    "bookings.fullname",
+    "bookings.mail",
+    "bookings.day",
+    "bookings.timetableId",
+    "timetables.startHour",
+    "timetables.endHour",
+    "bookings.createdAt",
+    "bookings.updatedAt",
+  ] satisfies ReadonlyArray<SelectExpression<DB, "bookings" | "timetables">>;
 
-  create(newBooking: CreateBooking): Promise<Booking> {
-    // Parse the string representation of the date
-    const bookingDate = { ...newBooking, day: new Date(newBooking.day) };
+  async create(newBooking: CreateBooking): Promise<Booking> {
+    const parsedBooking = { ...newBooking, day: new Date(newBooking.day) };
 
-    // Construct the fields to return, replacing the 'day' field with the SQL expression
-    const returnFields = this.DEFAULT_SELECT_FIELDS.map((field) =>
-      field === "day"
-        ? sql<string>`to_char(day, 'YYYY-MM-DD')`.as("day")
-        : field,
-    );
-
-    return this.db
+    const inserted = await this.db
       .insertInto("bookings")
-      .values(bookingDate)
-      .returning(returnFields)
+      .values(parsedBooking)
+      .returning(["bookings.id"])
       .executeTakeFirstOrThrow();
+
+    const insertedBooking = await this.db
+      .selectFrom("bookings")
+      .innerJoin("timetables", "bookings.timetableId", "timetables.id")
+      .select(
+        this.DEFAULT_SELECT_FIELDS.map((field) =>
+          field === "bookings.day"
+            ? sql<string>`to_char(bookings.day, 'YYYY-MM-DD')`.as("day")
+            : field,
+        ),
+      )
+      .where("bookings.id", "=", inserted.id)
+      .executeTakeFirstOrThrow();
+
+    return insertedBooking;
   }
 
   async findAll(
@@ -52,21 +62,35 @@ export class BookingDao implements IBookingRepository {
   ): Promise<PaginatedResult<Booking>> {
     let countQuery = this.db
       .selectFrom("bookings")
-      .select(({ fn }) => [fn.count<number>("id").as("count")]);
+      .select(({ fn }) => [fn.count<number>("bookings.id").as("count")]);
 
     // Construct the fields to return, replacing the 'day' field with the SQL expression
     const returnFields = this.DEFAULT_SELECT_FIELDS.map((field) =>
-      field === "day"
-        ? sql<string>`to_char(day, 'YYYY-MM-DD')`.as("day")
+      field === "bookings.day"
+        ? sql<string>`to_char(bookings.day, 'YYYY-MM-DD')`.as("day")
         : field,
     );
 
     let bookingsQuery = this.db
       .selectFrom("bookings")
+      .innerJoin("timetables", "timetables.id", "bookings.timetableId")
       .orderBy(buildSortBy<"bookings", Booking>(sortBy))
       .limit(pagination.limit)
       .offset(pagination.offset)
       .select(returnFields);
+
+    if (filterBy.timetableId) {
+      countQuery = countQuery.where(
+        "bookings.timetableId",
+        "=",
+        filterBy.timetableId,
+      );
+      bookingsQuery = bookingsQuery.where(
+        "bookings.timetableId",
+        "=",
+        filterBy.timetableId,
+      );
+    }
 
     if (filterBy.day) {
       countQuery = countQuery.where(
@@ -74,11 +98,11 @@ export class BookingDao implements IBookingRepository {
         "=",
         new Date(filterBy.day),
       );
-    }
-
-    if (filterBy.hour) {
-      countQuery = countQuery.where("bookings.hour", "=", filterBy.hour);
-      bookingsQuery = bookingsQuery.where("bookings.hour", "=", filterBy.hour);
+      bookingsQuery = bookingsQuery.where(
+        "bookings.day",
+        "=",
+        new Date(filterBy.day),
+      );
     }
 
     if (filterBy.mail) {
@@ -130,35 +154,57 @@ export class BookingDao implements IBookingRepository {
   async countBookingsForDayAndEmail(day: Date, mail: string): Promise<number> {
     const countResult = await this.db
       .selectFrom("bookings")
-      .select(({ fn }) => [fn.count<number>("id").as("count")])
-      .where((eb) => eb.and([eb("mail", "=", mail), eb("day", "=", day)]))
+      .select(({ fn }) => [fn.count<number>("bookings.id").as("count")])
+      .where((eb) =>
+        eb.and([eb("bookings.mail", "=", mail), eb("bookings.day", "=", day)]),
+      )
       .executeTakeFirst();
 
     return Number(countResult?.count ?? 0);
   }
 
-  async countBookingsForDayAndHour(day: Date, hour: string): Promise<number> {
+  async countBookingsForDayAndTimetableId(
+    day: Date,
+    timetableId: number,
+  ): Promise<number> {
     const countResult = await this.db
       .selectFrom("bookings")
-      .select(({ fn }) => [fn.count<number>("id").as("count")])
-      .where((eb) => eb.and([eb("day", "=", day), eb("hour", "=", hour)]))
+      .innerJoin("timetables", "timetables.id", "bookings.timetableId")
+      .select(({ fn }) => [fn.count<number>("bookings.id").as("count")])
+      .where((eb) =>
+        eb.and([
+          eb("bookings.day", "=", day),
+          eb("bookings.timetableId", "=", timetableId),
+        ]),
+      )
       .executeTakeFirst();
 
     return Number(countResult?.count ?? 0);
   }
 
-  delete(id: Booking["id"]): Promise<Booking | undefined> {
-    // Construct the fields to return, replacing the 'day' field with the SQL expression
-    const returnFields = this.DEFAULT_SELECT_FIELDS.map((field) =>
-      field === "day"
-        ? sql<string>`to_char(day, 'YYYY-MM-DD')`.as("day")
-        : field,
-    );
-
-    return this.db
-      .deleteFrom("bookings")
-      .where("id", "=", id)
-      .returning(returnFields)
+  async delete(id: Booking["id"]): Promise<Booking | undefined> {
+    const bookingToDelete = await this.db
+      .selectFrom("bookings")
+      .innerJoin("timetables", "bookings.timetableId", "timetables.id")
+      .select(
+        this.DEFAULT_SELECT_FIELDS.map((field) =>
+          field === "bookings.day"
+            ? sql<string>`to_char(bookings.day, 'YYYY-MM-DD')`.as("day")
+            : field,
+        ),
+      )
+      .where("bookings.id", "=", id)
       .executeTakeFirst();
+
+    if (!bookingToDelete) {
+      throw new NotFoundException(`Booking with id ${id} not found`);
+    }
+
+    await this.db
+      .deleteFrom("bookings")
+      .where("bookings.id", "=", id)
+      .executeTakeFirstOrThrow();
+
+    return bookingToDelete;
   }
 }
